requirements.txt
streamlit
boto3
botocore
python-dotenv
SpeechRecognition 
pyttsx3 
PyAudio
pytesseract 
pillow